refactor(api): update data types for certain Win32 entity fields

+++ src/main/java/io/github/eggy03/ferrumx/windows/entity/processor/Win32CacheMemory.java
@@ -52,7 +52,7 @@ public class Win32CacheMemory {
 
     @SerializedName("InstalledSize")
     @Nullable
-    Integer installedSize;
+    Long installedSize;
 
     @SerializedName("Associativity")
     @Nullable
