+++ src/main/java/io/github/eggy03/ferrumx/windows/entity/storage/Win32DiskPartition.java
@@ -84,7 +84,7 @@ public class Win32DiskPartition {
 
     @SerializedName("DiskIndex")
     @Nullable
-    Integer diskIndex;
+    Long diskIndex;
 
     @SerializedName("Size")
     @Nullable
